[MPS] expand the current export unit test for SDPA
Address review feedback: move the SDPA dynamic seq len export test from
test_mps.py to test_aot_inductor.py so it runs on all devices (cpu, gpu,
mps). Use F.scaled_dot_product_attention instead of the MPS-specific op.

Also clean up the meta registration for _scaled_dot_product_attention_math_for_mps:
inline the single-use sdpa_general_mps() helper and remove the dead
ensure_4d(key) call, both left over from the branching logic removed in #177620.

Authored with Claude.
